feat: add styles, upd right section design

+++ src/components/main/currentDate/CurrentDate.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 function CurrentDate() {
   const options = {
-    weekday: 'short',
+    weekday: 'long',
     month: 'long',
     day: 'numeric',
   };
